requirements.txt
aiogram==3.5.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
